--- a/include/cantera/kinetics/MultiRate.h
+++ b/include/cantera/kinetics/MultiRate.h
@@ -72,14 +72,9 @@ public:
         m_shared.update(T);
     }
 
-    virtual void update(double T, double P) override {
+    virtual void update(double T, double extra) override {
         // update common data once for each reaction type
-        m_shared.update(T, P);
-    }
-
-    virtual void update(double T, double P, double extra) override {
-        // update common data once for each reaction type
-        m_shared.update(T, P, extra);
+        m_shared.update(T, extra);
     }
 
     virtual bool update(const ThermoPhase& bulk, const Kinetics& kin) override {