--- a/include/cantera/kinetics/MultiRateBase.h
+++ b/include/cantera/kinetics/MultiRateBase.h
@@ -64,6 +64,12 @@ public:
     //! @param P  pressure [Pa]
     virtual void update(double T, double P) = 0;
 
+    //! Update common reaction rate data based on temperature and pressure
+    //! @param T  temperature [K]
+    //! @param P  pressure [Pa]
+    //! @param extra  extra parameter (depends on parameterization)
+    virtual void update(double T, double P, double extra) = 0;
+
     //! Update data common to reaction rates of a specific type
     //! @param bulk  object representing bulk phase
     //! @param kin  object representing kinetics