… include/cantera/kinetics/ReactionRate.h
… include/cantera/kinetics/RxnRates.h
… interfaces/cython/cantera/_cantera.pxd
… interfaces/cython/cantera/examples/kinetics/blowers_masel.py
… interfaces/cython/cantera/reaction.pyx
… interfaces/cython/cantera/test/test_kinetics.py
… interfaces/cython/cantera/test/test_reaction.py
… src/kinetics/Arrhenius.cpp
… src/kinetics/BulkKinetics.cpp
… src/kinetics/Custom.cpp
… src/kinetics/GasKinetics.cpp
… src/kinetics/ReactionData.cpp