<!-- Thanks for contributing code! Please include a description of your change and check your pull request against the list below. For further questions, refer to the contributing guide (https://github.com/Cantera/cantera/blob/main/CONTRIBUTING.md). -->

**Changes proposed in this pull request**

<!-- Provide a clear and concise description of changes and/or features introduced in this pull request. -->

- implement inheritance for `ReactionData`
- split `GasKinetics::updateROP` into multiple segments
- some preparatory work for a rebase of #1089

The PR further expands on ideas introduced in #1151

**If applicable, provide an example illustrating new features this pull request is introducing**

<!-- A minimal, complete, and reproducible example demonstrating features introduced by this pull request. See https://stackoverflow.com/help/minimal-reproducible-example for additional suggestions on how to create such an example. -->

With this PR:
```
$ PYTHONPATH=build/python python interfaces/cython/cantera/examples/kinetics/custom_reactions.py
Average time of 100 simulation runs for 'gri30.yaml' (CH4)
- New framework (YAML): 109.55 ms (T_final=2736.60)
- One Python reaction: 118.21 ms (T_final=2736.60) ... +7.90%
- Old framework (XML): 116.51 ms (T_final=2736.60) ... +6.35%
```
compared to current `main`:
```
$ PYTHONPATH=build/python python interfaces/cython/cantera/examples/kinetics/custom_reactions.py
Average time of 100 simulation runs for 'gri30.yaml' (CH4)
- New framework (YAML): 109.40 ms (T_final=2736.60)
- One Python reaction: 117.50 ms (T_final=2736.60) ... +7.41%
- Old framework (XML): 117.32 ms (T_final=2736.60) ... +7.24%
```
I ran the speed tests several times; there appears to be no noticeable impact of the code reorganization on performance. 

**Checklist**

- [ ] The pull request includes a clear description of this code change
- [ ] Commit messages have short titles and reference relevant issues
- [ ] Build passes (`scons build` & `scons test`) and unit tests address code coverage
- [ ] Style & formatting of contributed code follows [contributing guidelines](https://github.com/Cantera/cantera/blob/main/CONTRIBUTING.md)
- [ ] The pull request is ready for review