@ischoegl 2021-12-06
@speth ... these are some changes that are a preamble to #1089. While the `ReactionData` reorganization is a follow-up to your ideas in #1151, I am also including the split of `GasKinetics::updateROP` that #1089 introduces (the reorganization makes individual parts reusable). As this reshuffles things a bit, I wanted to post this now as this may create further conflicts with #1099 (@BangShiuh, sorry!).