@ischoegl 2021-12-12
@speth ... based on your prompt about the status of this PR, I had another look at what simplifications can be taken care of prior to #1089. It turns out that your ideas from #1101 and #1151 made a couple of other simplifications possible. 

* Now that `ReactionData` has become become a lot 'smarter' with #1151, I created an inheritance structure to eliminate redundant code.
* The `evalFromStruct` approach makes the `updateFromStruct` unnecessary, as after #1151 they are only called in combination, so the former can take care of everything
* I ended up replacing the 'buffering' approach used for `BlowersMaselRate` and `FalloffRate` by a more consistent update approach that adds these values to shared data used by the `ReactionRate._evaluator` mechanism introduced in #1101 
* I removed some of the service functions that I don't think will be needed, but kept some basic access. While some of your (highly useful!) `has_ddT` based templates are currently gone, they will come back with #1089