@ischoegl 2021-12-12
There was one additional simplification that I left on the table, which also made Python's `ReactionRate.__call__` API much cleaner. The PR should be good now.