@ischoegl 2021-12-13
@BangShiuh ... rebased and updated docstrings for the update mechanism (clarifying what is used when). Hope this helps with #1099.